Refactored conn_pair_01_to_03() to a separate module compat_utils.rs.
